CrudApi/Api/serializers.py
from rest_framework import serializers
from .models import *

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ['id', 'title', 'desc', 'timestamp', 'due_date', 'status', 'tags', 'tag_list']
